refactor: Update package.json and SearchBar component
This commit updates the package.json file to include the "@types/webpack-env" dependency with version "^1.18.5". It also modifies the SearchBar component to fetch file content from the "posts-submodule" directory and sets the items state variable with the fetched contents. This change improves the functionality and ensures that the correct file content is displayed in the SearchBar component.

Refactor package.json and SearchBar component to include "@types/webpack-env" dependency and fetch file content from "posts-submodule" directory.